Generate dummy config file for unsupported languages
The LaunguageHandlerBase performed a no-op for unknown languages
when generate_configs was called. This caused an error later
in the flow when the swagger gen tool is called with a config
file specified. To fix this, LaunguageHandlerBase now generates
a config with an empty json object so there is something to load.