Adding Post Processing for "Shadow Null" variables
The FlashArray REST API has some elements that are listed in the swagger spec as a `number`, but in fact accept either a number OR the empty string. When the empty string is passed, it means "unset this value". This is different from omitting or passing a `null` value for these variables, as the REST API interprets that as "don't change the current value".

We are calling these variables "Shadow Nullable", in that there is a mechanism to reset them that isn't supported by the Swagger as written. With the version of swagger currently used, there isn't a good way to add support for a value that can be either a number in a specific range, null, or the empty string. 

To address this issue, we've added a step in the Java post-processor which changes the internal data type of these variables from numeric to `Object`. We update the getters to cast back to the original type or `null`, and we add a `clearVariable` function which sets the internal value to the empty string.

---------

Co-authored-by: dulitha.gunasekera <dgunasekera@purestorage.com>